Added the deej functionality
Added the deej functionality in the qmk keyboards.

this uses virtser which is present on qmk.

this is tested with both RP2040 and STM32F401.
in RP2040 it works flawlessly and also works with vial.
in STM32F401 due to USB end point limitations, you can either turn on VIRTSER or other functionalities such as nkro, midi, bootmagic. but not both.